+++ Anarchy/Bridge/AnarchyBridge.cs
@@ -5,8 +5,8 @@
 namespace Anarchy.Bridge
 {
     using Anarchy.Components;
-    using Anarchy.Systems.AnarchyComponentsTool;
     using Anarchy.Systems.Common;
+    using Colossal.Entities;
     using Game.Tools;
     using Unity.Collections;
     using Unity.Entities;
@@ -88,14 +88,18 @@ namespace Anarchy.Bridge
                 return;
             }
 
+            EntityCommandBuffer buffer = new EntityCommandBuffer(Allocator.Temp);
             SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
             for (int i = 0; i <= entities.Length; i++)
             {
                 if (uiSystem.CheckOverridable(entities[i]))
                 {
-                    uiSystem.EntityManager.AddComponent<PreventOverride>(entities[i]);
+                    buffer.AddComponent<PreventOverride>(entities[i]);
                 }
             }
+
+            buffer.Playback(uiSystem.EntityManager);
+            buffer.Dispose();
         }
 
         /// <summary>
@@ -108,41 +112,166 @@ namespace Anarchy.Bridge
             {
                 return;
             }
-            /*
-            NativeArray<Entity> entities = entityQuery.ToEntityArray
-            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
-            if (uiSystem.CheckOverridable(instanceEntity))
-            {
-                uiSystem.EntityManager.AddComponent<PreventOverride>(instanceEntity);
-                return true;
-            }*/
 
-            return;
+            AddAnarchyComponent(entityQuery.ToEntityArray(Allocator.Temp));
         }
-        /*
+
         /// <summary>
-        /// Tries to add anarchy component to an instance entity.
+        /// Tries to add Transform Lock component to a entities in a query.
         /// </summary>
-        /// <param name="instanceEntity">Entity for an instance of an object. Must be overridable non-building static object.</param>
-        /// <param name="transform">Transform position and rotation to lock object.</param>
-        /// <returns>True if added. False if not.</returns>
-        public static bool TryAddTransformLockComponent(Entity instanceEntity, Game.Objects.Transform transform)
+        /// <param name="entityQuery">Entity query you want to try to add component to.</param>
+        public static void AddTransformLockComponent(EntityQuery entityQuery)
         {
-            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
-            if (uiSystem.CheckDisturbable(instanceEntity) &&
-               !uiSystem.EntityManager.HasComponent<TransformRecord>(instanceEntity))
+            if (entityQuery.IsEmptyIgnoreFilter)
             {
-                TransformRecord transformRecord = new TransformRecord()
-                {
-                    m_Position = transform.m_Position,
-                    m_Rotation = transform.m_Rotation,
-                };
-                uiSystem.EntityManager.AddComponent<TransformRecord>(instanceEntity);
-                uiSystem.EntityManager.SetComponentData(instanceEntity, transformRecord);
-                return true;
+                return;
             }
 
-            return false;
-        }*/
+            AddTransformLockComponent(entityQuery.ToEntityArray(Allocator.Temp));
+        }
+
+        /// <summary>
+        /// Tries to add Transform Lock component to entities in a native array.
+        /// </summary>c
+        /// <param name="entities">Native array of entities to try and add component to.</param>
+        public static void AddTransformLockComponent(NativeArray<Entity> entities)
+        {
+            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
+            EntityCommandBuffer buffer = new EntityCommandBuffer(Allocator.Temp);
+            for (int i = 0; i < entities.Length; i++)
+            {
+                if (uiSystem.CheckDisturbable(entities[i]) &&
+                   !uiSystem.EntityManager.HasComponent<TransformRecord>(entities[i]) &&
+                    uiSystem.EntityManager.TryGetComponent(entities[i], out Game.Objects.Transform transform))
+                {
+                    TransformRecord transformRecord = new TransformRecord()
+                    {
+                        m_Position = transform.m_Position,
+                        m_Rotation = transform.m_Rotation,
+                    };
+                    buffer.AddComponent<TransformRecord>(entities[i]);
+                    buffer.SetComponent(entities[i], transformRecord);
+                }
+            }
+
+            buffer.Playback(uiSystem.EntityManager);
+            buffer.Dispose();
+        }
+
+        /// <summary>
+        /// Removes Anarchy component from an entity if it has it.
+        /// </summary>
+        /// <param name="entity">Entity to check and remove component.</param>
+        public static void RemoveAnarchyComponent(Entity entity)
+        {
+            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
+            if (entity != Entity.Null &&
+                uiSystem.EntityManager.HasComponent<PreventOverride>(entity))
+            {
+                uiSystem.EntityManager.RemoveComponent<PreventOverride>(entity);
+            }
+        }
+
+        /// <summary>
+        /// Removes Transform Lock component from an entity if it has it.
+        /// </summary>
+        /// <param name="entity">Entity to check and remove component.</param>
+        public static void RemoveTransformLockComponent(Entity entity)
+        {
+            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
+            if (entity != Entity.Null &&
+                uiSystem.EntityManager.HasComponent<TransformRecord>(entity))
+            {
+                uiSystem.EntityManager.RemoveComponent<TransformRecord>(entity);
+            }
+        }
+
+        /// <summary>
+        /// Removes Anarchy component from an array of entities if they have it.
+        /// </summary>
+        /// <param name="entities">Entities to check and remove component.</param>
+        public static void RemoveAnarchyComponent(NativeArray<Entity> entities)
+        {
+            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
+            EntityCommandBuffer buffer = new EntityCommandBuffer(Allocator.Temp);
+            for (int i = 0; i < entities.Length; i++)
+            {
+                if (entities[i] != Entity.Null &&
+                    uiSystem.EntityManager.HasComponent<PreventOverride>(entities[i]))
+                {
+                    buffer.RemoveComponent<PreventOverride>(entities[i]);
+                }
+            }
+
+            buffer.Playback(uiSystem.EntityManager);
+            buffer.Dispose();
+        }
+
+        /// <summary>
+        /// Removes Transform Lock component from an array of entities if they have it.
+        /// </summary>
+        /// <param name="entities">Entities to check and remove component.</param>
+        public static void RemoveTransformLockComponent(NativeArray<Entity> entities)
+        {
+            SelectedInfoPanelTogglesSystem uiSystem = World.DefaultGameObjectInjectionWorld.GetOrCreateSystemManaged<SelectedInfoPanelTogglesSystem>();
+            EntityCommandBuffer buffer = new EntityCommandBuffer(Allocator.Temp);
+            for (int i = 0; i < entities.Length; i++)
+            {
+                if (entities[i] != Entity.Null &&
+                    uiSystem.EntityManager.HasComponent<TransformRecord>(entities[i]))
+                {
+                    buffer.RemoveComponent<TransformRecord>(entities[i]);
+                }
+            }
+
+            buffer.Playback(uiSystem.EntityManager);
+            buffer.Dispose();
+        }
+
+        /// <summary>
+        /// Removes Anarchy component from an query of entities if they have it.
+        /// </summary>
+        /// <param name="entityQuery">Query for entities to check and remove component.</param>
+        public static void RemoveAnarchyComponent(EntityQuery entityQuery)
+        {
+            if (entityQuery.IsEmptyIgnoreFilter)
+            {
+                return;
+            }
+
+            RemoveAnarchyComponent(entityQuery.ToEntityArray(Allocator.Temp));
+        }
+
+        /// <summary>
+        /// Removes Transform Lock component from an query of entities if they have it.
+        /// </summary>
+        /// <param name="entityQuery">Query for entities to check and remove component.</param>
+        public static void RemoveTransformLockComponent(EntityQuery entityQuery)
+        {
+            if (entityQuery.IsEmptyIgnoreFilter)
+            {
+                return;
+            }
+
+            RemoveTransformLockComponent(entityQuery.ToEntityArray(Allocator.Temp));
+        }
+
+        /// <summary>
+        /// Gets the component type for Anarchy Component.
+        /// </summary>
+        /// <returns>ComponentType for PreventOverride Component.</returns>
+        public static ComponentType GetAnarchyComponentType()
+        {
+            return ComponentType.ReadWrite<PreventOverride>();
+        }
+
+        /// <summary>
+        /// Gets the component type for Transform Lock.
+        /// </summary>
+        /// <returns>ComponentType for TransformRecord component.</returns>
+        public static ComponentType GetTransformLockComponentType()
+        {
+            return ComponentType.ReadWrite<TransformRecord>();
+        }
     }
 }