+++ Anarchy/Properties/Stable/ChangeLog.md
@@ -1,3 +1,7 @@
 ﻿## Patch V1.7.23
 *  Added bridge for other mods to register their tool systems.
 *  Added support for copying Anarchy and Transform Lock components using Move It Alpha's Copy/Paste.
+*  Added bridge for other mods to Add or Remove Anarchy or Transform Lock components to compatible game entities.
+*  Fixes to Flaming Chirper for game version 1.5.7f1. Thanks to Nullpinter for cleanest solution!
+*  Fix preview for Force Ground without Constant Slope.
+*  Updated Localization.